Separate types for contacts to own file
