+++ server/@types/index.d.ts
@@ -86,10 +86,6 @@ export interface DecoratedContact extends ContactSummaryResponse {
   }
 }
 
-export interface SelectableItem {
-  id?: number
-}
-
 export interface UiListItem {
   value: string
   text: string
@@ -107,11 +103,6 @@ export interface FormOption extends UiListItem {
   behaviour?: string
 }
 
-export interface SelectedFilterItem {
-  text: string
-  href: string
-}
-
 export interface UrlInfo {
   path: string
   fromPageId?: string
@@ -127,34 +118,6 @@ export interface FeatureFlagDefault {
 
 export type FeatureFlags = {
   [key: string]: boolean
-}
-
-export interface ContactTypeDecorated {
-  code: string
-  count: number
-  description: string
-}
-
-export interface ContactTypeGroupDecorated {
-  label: string
-  isGroupOpen: boolean
-  contactTypeCodes: {
-    value: string
-    description: string
-    html: string
-    count: number
-  }[]
-}
-
-export interface ContactHistoryFilters {
-  'dateFrom-day': string
-  'dateFrom-month': string
-  'dateFrom-year': string
-  'dateTo-day': string
-  'dateTo-month': string
-  'dateTo-year': string
-  contactTypes: string | string[]
-  searchFilters: string | string[]
 }
 
 export type PageId =
